+++ src/main/scala/org/tygus/suslik/certification/targets/coq/Coq.scala
@@ -1,7 +1,8 @@
 package org.tygus.suslik.certification.targets.coq
 
 import org.tygus.suslik.certification._
-import org.tygus.suslik.certification.targets.coq.Translation.TranslationException
+import org.tygus.suslik.certification.targets.coq.translation.Translation
+import org.tygus.suslik.certification.targets.coq.translation.Translation.TranslationException
 import org.tygus.suslik.language.Statements.Procedure
 import org.tygus.suslik.logic.Environment
