--- src/main/scala/org/tygus/suslik/certification/targets/coq/logic/package.scala
@@ -1,7 +1,0 @@
-package org.tygus.suslik.certification.targets.coq
-
-import org.tygus.suslik.certification.targets.coq.language.Expressions.CVar
-
-package object logic {
-  type ProofContext = Map[CVar, ProofContextItem]
-}
